+++ src/ULinkGame.Server/Actors/ActorRuntimeOptions.cs
@@ -1,0 +1,6 @@
+namespace ULinkGame.Server.Actors;
+
+public sealed class ActorRuntimeOptions
+{
+    public int MailboxCapacity { get; set; } = 4096;
+}