+++ src/ULinkGame.Server/Actors/IActor.cs
@@ -1,0 +1,5 @@
+namespace ULinkGame.Server.Actors;
+
+public interface IActor
+{
+}