+++ src/ULinkGame.Server/ULinkGameServerServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using ULinkGame.Server.Actors;
 using ULinkGame.Server.ReliablePush;
 using ULinkGame.Server.Sessions;
 
@@ -9,6 +10,7 @@ public static class ULinkGameServerServiceCollectionExtensions
 {
     public static IServiceCollection AddULinkGameServer(this IServiceCollection services)
     {
+        services.AddULinkGameServerActors();
         services.AddULinkGameServerSessions();
         services.AddULinkGameServerReliablePush();
         services.TryAddSingleton<IULinkGameServer, ULinkGameServer>();